requirement.txt
# Core dependencies
python-gedcom==2.0.0
networkx==3.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-Levenshtein==0.21.1
rapidfuzz==3.1.1

# PyTorch dependencies
torch==2.0.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
transformers==4.30.2

# MLX dependencies (Apple Silicon)
mlx==0.5.0
mlx-data==0.0.3

# Data processing
pyyaml==6.0
tqdm==4.65.0
matplotlib==3.7.1
seaborn==0.12.2

# Web API (for future)
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
joblib==1.3.1
ray==2.5.1  # For distributed deduplication
faiss-cpu==1.7.4  # For efficient similarity search